wait for the user confirmation before deleting

+++ src/utils/stories.ts
@@ -29,20 +29,22 @@ export const deleteStory = (
   store: Store,
   storyName: string,
 ) => {
-  try {
-    store.metadata.client.delete(storyName);
+  if (window.confirm(`Delete the '${storyName} story?'`)) {
+    try {
+      store.metadata.client.delete(storyName);
 
-    const withoutDeletedStory =
-      store.metadata.stories.filter(
-        (story) => story.name !== storyName,
+      const withoutDeletedStory =
+        store.metadata.stories.filter(
+          (story) => story.name !== storyName,
+        );
+
+      store.setStories(withoutDeletedStory);
+    } catch (e) {
+      alert(
+        `Could not delete story ${storyName}. See console for details.`,
       );
-
-    store.setStories(withoutDeletedStory);
-  } catch (e) {
-    alert(
-      `Could not delete story ${storyName}. See console for details.`,
-    );
-    console.error(e);
+      console.error(e);
+    }
   }
 };
